Camera Viewport typecast to int

+++ core/src/net/mueller_martin/turirun/WorldController.java
@@ -117,7 +117,7 @@ public class WorldController {
         float cameraHalfHeight = cam.viewportHeight * .5f;
 
         // Move camera after player as normal
-        CameraHelper.instance.camera.position.set(controller.character.currentPosition.x,controller.character.currentPosition.y,0);
+        CameraHelper.instance.camera.position.set((int)controller.character.currentPosition.x,(int)controller.character.currentPosition.y,0);
 
         float cameraLeft = cam.position.x - cameraHalfWidth;
         float cameraRight = cam.position.x + cameraHalfWidth;
